struct vector redefined as double
pretty much changed all the files. functions switched to double version(without f)

+++ collision code/mesh.h
@@ -8,8 +8,8 @@ ParticleMesh* create_particle_mesh(void);
 void destroy_particle_mesh(ParticleMesh *pm);
 
 // Grid operations
-float*** allocate_3d_array(int n);
-void free_3d_array(float ***array, int n);
+double*** allocate_3d_array(int n);
+void free_3d_array(double ***array, int n);
 
 // Mass assignment
 void assign_mass_cic(ParticleMesh *pm, ParticleSystem *sys);
